Allow to build windows binaries
